210126 pull request

+++ src/main/java/org/zerock/domain/CategoryVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CategoryVO {
 	
-	private int category_code;
+	private long category_code;
 	private String category_name;
 
 }
